coverage.ansi: Deal with projects that have fail-under set
Say you have a .coveragerc with [report] fail-under=95 set. Even
coverage json or coverage xml will exit 1 and print to stderr a warning
about insufficient coverage, which they really shouldn't. We're not
going to bother trying to parsing the still emitted and valid JSON in
that case, but just prevent it in the first place.